src/tasks/task3/notebooks/Minto_agent_langchain_multiple_crew/requirements.txt
crewai==0.30.11
langchain==0.2.2
langchain_community==0.2.3
langchain_core==0.2.4
langchain_openai==0.1.8
python-dotenv==1.0.1
streamlit==1.33.0
